# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem+json 错误体——二期新端点的统一错误形状。\n\n与一期 `ErrorDetail`（`{\"detail\"}`）并存：一期端点不返工，前端读 detail 兼容两者。\n扩展字段（occupier / 冲突清单等）由各错误在响应里按需附加，不进本模型。",
        "properties": {
          "detail": {
            "description": "中文可操作消息——下一步该做什么",
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "description": "HTTP 状态码（与响应状态一致）",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "人读的短语概括",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "机器可读的错误类别 slug（如 task-not-found）",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RetryListRequest": {
        "description": "重试列表加入的请求体：条目数组（素材主干）。",
        "properties": {
          "items": {
            "description": "条目清单（素材主干，如 cat_001）",
            "items": {
              "type": "string"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RetryListRequest",
        "type": "object"
      },
      "RetryListView": {
        "description": "重试列表现状（加入 / 移出 / 清空都返回全量名单，前端以响应为准）。",
        "properties": {
          "id": {
            "description": "批次标识（sN 形式）",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "description": "当前重试名单（名单顺序即重试顺序）",
            "items": {
              "type": "string"
            },
            "title": "Items",
            "type": "array"
          },
          "seq": {
            "description": "序号",
            "title": "Seq",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "seq",
          "items"
        ],
        "title": "RetryListView",
        "type": "object"
      },
      "RunAccepted": {
        "description": "跑批受理响应（202）：后台线程已受理，进度走 current / SSE。",
        "properties": {
          "run_id": {
            "description": "运行 id（.dsf/runs/ 下的目录名）",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "RunAccepted",
        "type": "object"
      },
...
      "RunHistoryView": {
        "description": "最近运行的磁盘摘要及可定位的日志绝对路径。",
        "properties": {
          "items_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items Path"
          },
          "log_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Path"
          },
          "record": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunRecord"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "record",
          "log_path",
          "items_path"
        ],
        "title": "RunHistoryView",
        "type": "object"
      },
...
      "RunStartRequest": {
        "description": "启动跑批的请求体。",
        "properties": {
          "items": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "retry 模式的可选明确条目；省略时使用整份重试名单",
            "title": "Items"
          },
          "mode": {
            "description": "full = 全量打未完成的条目；retry = 只打重试列表快照",
            "enum": [
              "full",
              "retry"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "mode"
        ],
        "title": "RunStartRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "AI 打标工具（发图 + 指令产出 caption，支持迭代改写）",
    "title": "Dataset Factory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workdirs/{wid}/batches/{sN}/retry-list": {
      "delete": {
        "description": "整体清空本批次的重试列表（其他批次的名单不动），返回空名单。",
        "operationId": "clear_batch_retry_list_api_workdirs__wid__batches__sN__retry_list_delete",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "wid 或批次不存在（workdir-not-found / batch-not-found）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Batch Retry List",
        "tags": [
          "跑批"
        ]
      },
      "post": {
        "description": "把条目加入重试列表（幂等去重），返回当前名单（名单顺序即重试顺序）。\n\n只收「已完成」与「可重试的未完成」条目——排队中无需重试、缺失要\n先补素材、不可重试失败要先解决格式问题；资格用当刻的条目视图现判。改动经\nmutate_state 在状态锁内完成；运行期写入照常受理（本次运行按启动时的快照执行）。",
        "operationId": "add_batch_retry_list_api_workdirs__wid__batches__sN__retry_list_post",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "wid 或批次不存在（workdir-not-found / batch-not-found）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "有不可入列的条目（problem+json: retry-item-not-eligible），rejections 扩展字段带逐条原因；整体拒绝、不做部分入列"
          }
        },
        "summary": "Add Batch Retry List",
        "tags": [
          "跑批"
        ]
      }
    },
    "/api/workdirs/{wid}/batches/{sN}/retry-list/{item}": {
      "delete": {
        "description": "把一个条目移出重试列表（幂等：不在名单里时原样返回），返回当前名单。",
        "operationId": "remove_batch_retry_list_item_api_workdirs__wid__batches__sN__retry_list__item__delete",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item",
            "required": true,
            "schema": {
              "title": "Item",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "wid 或批次不存在（workdir-not-found / batch-not-found）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Batch Retry List Item",
        "tags": [
          "跑批"
        ]
      }
    },
    "/api/workdirs/{wid}/batches/{sN}/runs": {
      "post": {
        "description": "受理一次跑批：同步预检 + 登记注册表 + 后台线程执行，202 立即返回。\n\n预检在受理之前完成（批次存在 / 活跃 / 快照可读 / 客户端可装配），把能在\n请求内确定的错误同步报给客户端；真正的运行锁在后台线程里抢——跨进程占用\n（如 CLI 正在跑同一工作目录）在受理后才暴露，失败原因经 SSE / current 呈现\n（进度快照 status=failed + 一条 failed 的 run-finished 事件）。",
        "operationId": "start_run_api_workdirs__wid__batches__sN__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "wid / 批次 / 快照不存在（problem+json）"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "批次已停用或工作目录已有跑批在运行（problem+json: batch-inactive / run-occupied）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Run",
        "tags": [
          "跑批"
        ]
      }
    },
    "/api/workdirs/{wid}/batches/{sN}/runs/current": {
      "get": {
        "description": "当前运行进度快照（轮询用；SSE 断线重连后的全量刷新同款数据）。\n\n没有进行中的跑批时返回 200 + null（2026-09-21 审计定案 L3/B6）：空闲轮询是\n前端的合法问询，报 404 会把日志刷成错误流、把真错误淹掉。",
        "operationId": "current_run_api_workdirs__wid__batches__sN__runs_current_get",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RunStatusView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Current Run Api Workdirs  Wid  Batches  Sn  Runs Current Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "wid 或批次不存在（workdir-not-found / batch-not-found）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Current Run",
        "tags": [
          "跑批"
        ]
      }
    },
    "/api/workdirs/{wid}/batches/{sN}/runs/latest": {
      "get": {
        "description": "按批次回读最近一次磁盘记录；尚无运行时返回空摘要。",
        "operationId": "latest_run_api_workdirs__wid__batches__sN__runs_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunHistoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Run",
        "tags": [
          "跑批"
        ]
      }
    },
    "/api/workdirs/{wid}/batches/{sN}/runs/stop": {
      "post": {
        "description": "请求停止当前跑批（协作取消）：置位信号即返回，当前条目在安全点停下。",
        "operationId": "stop_run_api_workdirs__wid__batches__sN__runs_stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "该批次当前没有进行中的跑批（problem+json: run-not-active）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Run",
        "tags": [
          "跑批"
        ]
      }
    },
    "/api/workdirs/{wid}/batches/{sN}/runs/stream": {
      "get": {
        "description": "SSE 业务事件流：run-started → item-updated… → run-finished（收到终态即关流）。\n\n订阅之前已发生的事件不补发（前端断线约定：重连先全量拉条目视图刷新界面）；\n事件经线程安全队列从跑批线程转发到流。",
        "operationId": "stream_run_api_workdirs__wid__batches__sN__runs_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "该批次当前没有进行中的跑批（problem+json: run-not-active）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run",
        "tags": [
          "跑批"
        ]
      }
    },
    "/api/workdirs/{wid}/batches/{sN}/runs/{run_id}/text": {
      "get": {
        "description": "指定运行查看日志，打开后不因新运行出现而切换文件。",
        "operationId": "run_text_api_workdirs__wid__batches__sN__runs__run_id__text_get",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sN",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "file",
            "required": false,
            "schema": {
              "default": "run.log",
              "enum": [
                "run.log",
                "items.jsonl"
              ],
              "title": "File",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunTextView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Text",
        "tags": [
          "跑批"
        ]
      }
    },
...
  }
}